changed matrix to use double instead of int

+++ WeatherMatrix/Matrix.cpp
@@ -5,10 +5,10 @@ Matrix::Matrix(int l, int h)
 	_length = l;
 	_height = h;
 
-	_myArr = new int*[_height];
+	_myArr = new double*[_height];
 
 	for (int i = 0; i < _height; i++)
-		_myArr[i] = new int[_length];
+		_myArr[i] = new double[_length];
 }
 
 Matrix::~Matrix()
@@ -20,10 +20,10 @@ Matrix::~Matrix()
 
 void Matrix::operator = (const Matrix & copyFrom)
 {
-	_myArr = new int*[_height];
+	_myArr = new double*[_height];
 
 	for (int i = 0; i < _height; i++)
-		_myArr[i] = new int[_length];
+		_myArr[i] = new double[_length];
 
 	_length = copyFrom._length;
 	_height = copyFrom._height;
